backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pillow==10.4.0
pytesseract==0.3.13
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
tesseract
numpy
pandas 
matplotlib
sklearn
python-dotenv==1.0.1
requests==2.32.3
pdfplumber==0.11.4


